Move getCurrentAppearence out
