+++ packages/theming/theming-utils/.eslintrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+  extends: [
+    "@uifabricshared/eslint-config-rules"
+  ],
+  rules: {
+    '@typescript-eslint/no-var-requires': 0,
+  }
+}
