+++ packages/theming/theming-utils/api-extractor.json
@@ -1,0 +1,3 @@
+{
+  "extends": "@uifabricshared/build-native/api-extractor.json"
+}
