+++ packages/theming/theming-utils/babel.config.js
@@ -1,0 +1,1 @@
+module.exports = require('@uifabricshared/build-native/babel.config');
