+++ packages/theming/theming-utils/jest.config.js
@@ -1,0 +1,2 @@
+const { configureReactNativeJest } = require('@uifabricshared/build-native');
+module.exports = configureReactNativeJest('android');
