+++ packages/theming/theming-utils/just.config.js
@@ -1,0 +1,3 @@
+const { preset } = require('@uifabricshared/build-native');
+
+preset();
